+++ apps/server/CHANGELOG.md
@@ -1,5 +1,30 @@
 # @stackspulse/server
 
+## 0.9.2
+
+### Patch Changes
+
+- 6e4b72c: Add /health route.
+
+## 0.9.1
+
+## 0.9.0
+
+### Patch Changes
+
+- d0f542a: Move /api/root/upsert-dapps to api server.
+
+## 0.8.0
+
+### Minor Changes
+
+- cc29346: Create new `/transactions/stats` route.
+
+### Patch Changes
+
+- 20d47b2: Add Fly config.
+- f0129c3: Add cors config.
+
 ## 0.7.2
 
 ### Patch Changes